requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==5.0.7
django-cors-headers==4.4.0
django-seed==0.3.1
django-seeding==1.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
Faker==26.0.0
filelock==3.15.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
numpy==2.0.0
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
six==1.16.0
sqlparse==0.5.0
tenacity==9.0.0
tldextract==5.1.2
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.event==5.0
zope.interface==6.4.post2
